Day 5
Completed day 5, this one involved a lot of nths and unwrapping in order
to parse the input. The input is very easy graphically to understand but
a little more annoying for parsing than simple delineation. It was also
the day of off by one errors as a result.

Definitely could have a cleaner implementation. First use of a custom
struct and it made a fair amount of sense to. Learned that `split_at`
doesn't consume the char at the split index, so have to discard that
after. Too much working with state in the solution. Still, this was an
interesting one!