Day 6
This one was pretty simple to implement, although I don't know how
rust-y my code is. I found fasterthanlime's series on AoC today, so I'll
be reading those to compare how my code could be better!

I write this one without the marker_len to begin with, but it was very
easy to slot in to fix part two of the puzzle, just extending the
marker length from 4 to 14. Fun little puzzle. The string function for
removing the 0th value is O(n), which makes the get_marker func O(n*k)
where n is the input length and k is the length of the unique character
string. It could be improved by using a deque instead of a string there.
Not too sure about the is_marker function, it'll instantiate a HashSet
each time which is a little more expensive computationally, but in terms
of verifying if a string contains only unique chars it is O(n). Actually
that makes this solution O(n*k*k) - switching to a deque (giving O(1)
string editing) would make this O(n*k) overall.