Day 7 failure
Completely failed day 7! Got hung up with parsing and couldn't think of
how to deal with the data structures. Ended up entirely cribbing off the
work of fasterthanlime - at least learning a lot about how to deal with
recursive data structures in rust in the process. id_tree seems a very
handy library for doing this sort of thing.

My initial thought was to use recursive functions and just return from
there, but didn't get to that