Day 10
This is a really fun puzzle idea! Loved the drawing out of pixels part.

As seems to be traditional for me so far, the parsinmg took the longest
part. Even though this is incredibly simple, I wanted to give using
`nom` a go again, but that backfired as I struggled to parse the add
instruction. Found it very difficult to debug, couldn't get it workjing
or figure out where I was going wrong. In the end I went back to very
simple split on whitespace parsing, which worked absolutely fine!

My logic for thje instruction handling is pretty simple. I created a
deque for the instructions, and a temporary register for the current
instruction. If the current instruction is empty I load a new one into
it (noops count as empty instructions)/. If it's not empty I add the
number to the X register.

Logic for drawing the pixels is similarly simple. Good fun puzzle!