Day 11 complete
This was so fun! I really enjoyed the part two problem. My immediate
reaction was to just try increasing the size of the numbers, going for
bigger and bigger ints. Really love how easy Rust makes that
experimentation and how FAST it is, immediate turn around. I feel for
the people today using python for this that probably sat there waiting
for their code to run as it created larger and larger numbers until the
stack overflowed.

After stepping away for a second I realised the clue to find some other
way to decrease your worry. Figured it out in the shower, really
satisfying to realise there's a mathematical solution. Here it was to
find a common multiple of the monkeys divisors (I just multiplied them
as I didn't want to code a lowest common multiplier function). That
meant the behaviour of the logic checks is consistent as the checks
operate in a field (or is it a ring, not perfect at my number theory).

Great puzzle, as always parsing took me an age, starting to get their
with nom parsing though! My method is now lots of debug statements while
I'm coding it and to just grab whatever I can with let statements. It's
not the cleanest but it works for these very short parsing problems.