Day 13 Complete
This was a little bit of a pig with the parsing and data structure. Took
me some time overnight to figure out what data structure was actually
possible and would work for the packet. Again took me some time parsing
with `nom`, but once I figured out how to do a neat recursive parse it
was incredibly satisfying and concise.

The ordering implementation is less concise due to the unboxing of the
variables from the enum, some times I needed them still in the container
and sometimes not, but it was again very satisfying to get working. Had
a funny bug where I was getting the wrong order because I was deriving
PartialOrd, but properly implementing that for the Data type fixed it.

Great puzzle all round!