Day 14
I love rust because even when it's pretty unoptimized code it still runs
very quickly. I took a really straightforward solution here, represent
the grid as a flat array of bools with methods for indexing in by (x,y).
I parsed the lines and drew them in to the grid, then added a method for
dropping a single grain of sand. The dropping method is particularly
unoptimized as it is worst case O(h) where h is the height of the array
to come to rest. If I wanted to optimize this I would memoize with an
array that stores the highest point of each stage in the cave, and then
on changing columns the sand would 'drop' down to that point.

But at the moment the code is nice and clear, and still completes in
under 6ms system time on a release build.  Really love how ergonomic rust
feels for this type of thing too.

Ran into a memory bug at one point when I tried making it an array of
chars, forgetting that each char takes up 4 bytes. 4*500kb was enough to
overflow the stack, could have Boxed it but the code obviously didn't
need that.